# perf(select): Use byte records instead of string records