```
 Benchmark 1: ./pica-v0.25 select -s -H 'ppn,gnd_id,label' '003@.0,041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "O" && 041A/*.9?' -o out.csv
  Time (mean ± σ):     261.160 s ±  2.054 s    [User: 240.933 s, System: 18.878 s]
  Range (min … max):   259.707 s … 262.612 s    2 runs

Benchmark 2: ./pica-main select -s -H 'ppn,gnd_id,label' '003@.0,041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "O" && 041A/*.9?' -o out.csv
  Time (mean ± σ):     243.547 s ±  1.746 s    [User: 223.634 s, System: 18.478 s]
  Range (min … max):   242.312 s … 244.782 s    2 runs

Summary
  './pica-main select -s -H 'ppn,gnd_id,label' '003@.0,041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "O" && 041A/*.9?' -o out.csv' ran
    1.07 ± 0.01 times faster than './pica-v0.25 select -s -H 'ppn,gnd_id,label' '003@.0,041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "O" && 041A/*.9?' -o out.csv'
```